pyproject.toml
[tool.poetry]
name = "seesaw"
version = "0.2.0"
description = "Seesaw"
authors = ["Oscar Moll <orm@csail.mit.edu>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
torch = "^1.11.0"
torchvision = "^0.12.0"
pytorch-lightning = "^1.6.3"
ray = {extras = ["tune"], version = "^1.12.0"}
Pillow = "^9.1.0"
pyroaring = "^0.3.3"
pyarrow = "^8.0.0"
annoy = "^1.17.0"
tqdm = "^4.64.0"
fastapi = "^0.77.1"
pydantic = "^1.9.0"
scikit-learn = "^1.1.0"
pandas = "^1.4.2"
transformers = "^4.19.0"
datasets = "^2.2.1"
clip = {git = "https://github.com/openai/CLIP.git", rev = "e58d49454c9"}
ftfy = "^6.1.1"

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"
plotnine = "^0.8.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
